admin-dashboard/requirements.txt
# requirements.txt
streamlit
supabase
pandas
plotly
python-dotenv
numpy
requests